+++ Sming/Libraries/Timezone/README.rst
@@ -1,0 +1,8 @@
+Timezone
+========
+
+Arduino library to support local/UTC time conversions using rules
+
+See :sample:`SystemClock_NTP` for example usage.
+
+Port of https://github.com/JChristensen/Timezone for Sming.
